Add missing features and stablize the app

+++ libs/backend/db-access/src/services/project/project-member.service.ts
@@ -1,6 +1,10 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import {
   ProjectMember,
   ProjectMemberDocument,
@@ -13,12 +17,15 @@ import {
   ListResponse,
 } from '@tma/shared/api-model';
 import { ProjectDocument } from '../../entities/project.entity';
+import { User, UserDocument } from '../../entities/user';
 
 @Injectable()
 export class ProjectMemberService {
   constructor(
     @InjectModel(ProjectMember.name)
-    private readonly model: Model<ProjectMemberDocument>
+    private readonly model: Model<ProjectMemberDocument>,
+    @InjectModel(User.name)
+    private readonly userModel: Model<UserDocument>
   ) {}
   async getMemberProjects(
     userId: string,
@@ -27,19 +34,40 @@ export class ProjectMemberService {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     filter: any
   ): Promise<ListResponse<ProjectDocument>> {
-    const filters = { userId };
+    const userObjectId = new Types.ObjectId(userId);
+    const filters = { user: userObjectId };
     if (filter) {
       Object.assign(filters, filter);
     }
     const count = await this.model.countDocuments(filters).exec();
+    if (count == 0) {
+      return Promise.resolve({
+        items: [],
+        totalItems: count,
+      });
+    }
     const list = await this.model
       .find(filters)
       .skip(offset)
       .limit(limit)
-      .populate('project')
+      .lean()
+      .populate({
+        path: 'project',
+        populate: [
+          {
+            path: 'createdBy',
+            select: 'name email username _id',
+          },
+          {
+            path: 'modifiedBy',
+            select: 'name email username _id',
+          },
+        ],
+      })
       .exec();
     const docs: ProjectDocument[] = list.map((item) => {
-      return item.project;
+      const prj = item.project;
+      return prj;
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     }) as any;
     const response: ListResponse<ProjectDocument> = {
@@ -48,7 +76,7 @@ export class ProjectMemberService {
     };
     return Promise.resolve(response);
   }
-  async getProjectMembers(projectId: string): Promise<ProjectMemberDTO> {
+  async getProjectMembers(projectId: string): Promise<ProjectMemberDTO[]> {
     const list = await this.model
       .find({ project: projectId })
       .populate('user')
@@ -82,6 +110,14 @@ export class ProjectMemberService {
     memberId: string
   ): Promise<GenericSuccessResponse> {
     this.checkProjectAccess(projectId, currentUser);
+    const members = await this.getProjectMembers(projectId);
+    if (
+      !members.some(
+        (member) => member.role === Role.Owner && member.user.id !== memberId
+      )
+    ) {
+      throw new BadRequestException('Project atleast needs one owner.');
+    }
     await this.model.deleteOne({ project: projectId, user: memberId }).exec();
     return Promise.resolve({ result: 'success' });
   }
@@ -92,12 +128,25 @@ export class ProjectMemberService {
     member: AddProjectMemberDTO
   ): Promise<GenericSuccessResponse> {
     this.checkProjectAccess(projectId, currentUser);
+    const uId = new Types.ObjectId(member.userId);
+    const pId = new Types.ObjectId(projectId);
+    if (member.role == Role.Member) {
+      const members = await this.getProjectMembers(projectId);
+      if (
+        !members.some(
+          (member1) =>
+            member1.role === Role.Owner && member1.user.id !== member.userId
+        )
+      ) {
+        throw new BadRequestException('Project atleast needs one owner.');
+      }
+    }
     await this.model
       .findOneAndUpdate(
-        { project: projectId, user: member.userId },
+        { project: pId, user: uId },
         {
-          project: projectId,
-          user: member.userId,
+          project: pId,
+          user: uId,
           role: member.role,
         },
         {
@@ -114,8 +163,10 @@ export class ProjectMemberService {
     userId: string,
     checkRole = true
   ): Promise<boolean> {
+    const uId = new Types.ObjectId(userId);
+    const pId = new Types.ObjectId(projectId);
     const projectMember = await this.model
-      .findOne({ project: projectId, user: userId })
+      .findOne({ project: pId, user: uId })
       .exec();
     if (projectMember == null) {
       throw new ForbiddenException(`You don't have aacess to the project.`);
@@ -131,6 +182,10 @@ export class ProjectMemberService {
     projectId: string
   ): Promise<GenericSuccessResponse> {
     await this.model.deleteMany({ project: projectId }).exec();
+    return Promise.resolve({ result: 'success' });
+  }
+  async deleteUserEntries(userId: string): Promise<GenericSuccessResponse> {
+    await this.model.deleteMany({ user: userId }).exec();
     return Promise.resolve({ result: 'success' });
   }
 }
